Volt-Route-Backend/requirements.txt
Django>=5.1.3
djangorestframework>=3.14
psycopg2-binary>=2.9
# Optional helper for env management (not required)
python-dotenv>=1.0
